backend/requirements.txt
Flask
flask-cors
psycopg2
cloudinary
python-dotenv
werkzeug
pyjwt
